config: one Mode enum for every conv kernel knob, typed setHint(Hint, Mode)
Replace WinogradMode + TuningLevel with a single Mode enum that names every conv
kernel-selection value (Auto/On/Off, NoTune/Fast/Thorough, TiledGemm..SubgroupGemm,
F23/F43, DirectAuto/RegisterTiled/LdsHalo). Add a typed setHint(Hint, Mode)
overload so callers pass the enum directly, no int cast:

    cfg.setHint(Hint::Winograd,        Mode::Off);
    cfg.setHint(Hint::WinogradVariant, Mode::SubgroupGemm);

Expose the variant/unit/direct hints in the benchmark config (winogradVariant /
winogradUnit / directConv3x3) and forward them through run.py, so every conv
kernel knob is settable from a config, not just programmatically.

+++ examples/classify.cpp
@@ -80,9 +80,9 @@ int main(int argc, char **argv) {
     cfg.cacheDir  = argval(argc, argv, "--cache", cfg.cacheDir.c_str());
     {
         std::string w = argval(argc, argv, "--winograd", "auto"); // auto|on|off
-        cfg.setHint(Hint::Winograd, (int) winogradFromStr(w));
+        cfg.setHint(Hint::Winograd, winogradFromStr(w));
         std::string t = argval(argc, argv, "--tuning", "fast"); // off|fast|thorough
-        cfg.setHint(Hint::Tuning, (int) tuningFromStr(t));
+        cfg.setHint(Hint::Tuning, tuningFromStr(t));
         // Advanced hints: force Winograd unit / variant for research.
         int wu = atoi(argval(argc, argv, "--wino-unit", "0")); // 0=auto, 4=force F(4,3)
         if (wu)
